Edit : Event Feed Cell
혜리님이 작업하신 부분 받아서 오토레이아웃 수정하고 쉐도우 뷰 추가했습니다!
